Restructure the components and styling
Restructure the files by high order components and associate their styles with them. This way we can also add in testing easily in the future.
