+++ src/App.css
@@ -16,112 +16,3 @@ body {
 .search-bar input{
   width: 500px;
 }
-
-.playbook-card {
-  height: 287px;
-  width: 284px;
-  text-align: center;
-}
-
-
-.cart-contents {
-  width: 640px;
-  background-color: white;
-}
-
-.cart-item {
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-  height: 79px;
-  padding-top: 10px;
-  padding-bottom: 10px;
-  border: solid #dedef0 1px;
-}
-
-.cart-item .cart-item--actions{
-  display: flex;
-  flex-direction: column;
-}
-
-.cart-item .cart-item--image{
-  margin-left: 10px;
-}
-
-.cart-item .cart-item--price{
-  margin-right: 10px;
-}
-
-.my-cart-page-contents {
-  display: flex;
-}
-
-.my-cart-page-contents .my-cart-page__right{
-  margin-left: 40px;
-}
-
-.my-cart-page-contents .my-cart-page__right .cart-checkout-action{
-  margin-top: 20px;
-}
-
-.splash-page {
-  margin-left: -10%;
-  margin-right: -10%;
-  margin-top: -20px;
-}
-
-.splash-title {
-  text-align: center;
-}
-
-.splash-content-container-1 {
-  font-size: 150px;
-  font-weight: 700;
-  text-align: center;
-  width: 100%;
-  height: 600px;
-  display: flex;
-  justify-content: center;
-  flex-direction: column;
-  background-image: url("./images/stock_photo1.jpg");
-  background-repeat: no-repeat;
-  background-size: cover;
-  background-position: center center;
-}
-
-.stock-photo-container {
-  height: 550px;
-  width: 100%;
-  text-align: center;
-}
-
-.stock-photo {
-  width: 100%;
-  height: 100%;
-  overflow: hidden;
-  object-fit: cover;
-}
-
-.tri-content-col {
-  width: 33.33%;
-  height: 400px;
-  float: left;
-  text-align: center;
-}
-
-.tri-content-col.even:nth-child(even) {
-  background-color: white;
-}
-.tri-content-col.odd:nth-child(odd) {
-  background-color: white;
-}
-
-.splash-footer-container {
-  height: 400px;
-  width: 100%;
-  text-align: center;
-}
-
-.splash-footer-content {
-  padding: 20px;
-}
