+++ src/App.js
@@ -1,6 +1,5 @@
 import React, { Component } from 'react';
 import './App.css';
-import '../src/styles/components/Header.css';
 import AppRouter from './routers/AppRouter';
 
 // class App extends Component {
