+++ src/components/PlaybookCard/PlaybookCard.css
@@ -1,0 +1,5 @@
+.playbook-card {
+  height: 287px;
+  width: 284px;
+  text-align: center;
+}
