+++ src/components/PlaybookCard/PlaybookCard.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Card } from 'semantic-ui-react'
-import { Link } from 'react-router-dom';
+import "./PlaybookCard.css"
 
 const PlaybookCard = ({ name, id }) => (
     <Card
