+++ src/components/PlaybookDetailsPage/ShoppingTile.js
