--- src/styles/components/PlaybookCard.css
@@ -1,4 +1,0 @@
-.playbook-card {
-    height: 200px;
-    width: 400px;
-}
